+++ extend.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Password Strength Checker Extension for Flarum.
+ * Password Strength Indicator for Flarum.
  *
  * LICENSE: For the full copyright and license information,
  * please view the LICENSE.md file that was distributed
@@ -11,7 +11,7 @@
  * @author     Hasan Özbey <hasanoozbey@gmail.com>
  * @copyright  2020
  * @license    The MIT License
- * @version    Release: 0.1.0
+ * @version    Release: 1.0.0
  * @link       https://github.com/the-turk/password-strength
  */
 
@@ -19,16 +19,16 @@ namespace TheTurk\PasswordStrength;
 
 use Flarum\Extend;
 use Illuminate\Contracts\Events\Dispatcher;
-use TheTurk\PasswordStrength;
 
 return [
     (new Extend\Frontend('forum'))
+        ->css(__DIR__ . '/less/forum.less')
         ->js(__DIR__.'/js/dist/forum.js'),
     (new Extend\Frontend('admin'))
-        ->css(__DIR__ . '/resources/less/admin/settingsModal.less')
+        ->css(__DIR__ . '/less/admin.less')
         ->js(__DIR__ . '/js/dist/admin.js'),
     (new Extend\Locales(__DIR__ . '/locale')),
     function (Dispatcher $events) {
-        $events->subscribe(PasswordStrength\Listeners\LoadSettings::class);
+        $events->subscribe(Listeners\LoadSettings::class);
     },
 ];